pyproject.toml
[project]
name = "lifetracker"
version = "0.1.0"
description = "Vibe Check — Personal Like/Dislike Activity Tracker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot", "core", "api"]
